Backend/requirements.txt
flask
face-recognition
numpy
pillow
opencv-python
requests
io